+++ c++/triqs/gfs/functions/imtime.cpp
@@ -31,7 +31,7 @@ namespace triqs {
     gf<mesh::imtime> rebinning_tau(gf_const_view<mesh::imtime> const &g, size_t new_n_tau) {
 
       auto const &old_m = g.mesh();
-      gf<mesh::imtime> new_gf{{old_m.domain().beta, old_m.domain().statistic, new_n_tau}, g.target_shape()};
+      gf<mesh::imtime> new_gf{{old_m.domain().beta, old_m.domain().statistic, static_cast<long>(new_n_tau)}, g.target_shape()};
       auto const &new_m = new_gf.mesh();
       new_gf.data()()   = 0;
       long prev_index   = 0;
